+++ src/Components/Home Page/ViewImg.jsx
@@ -1,0 +1,12 @@
+export function ViewImg() {
+    return <section style={{ textAlign: "center" }}>
+    <img
+      src="//cdn.shopify.com/s/files/1/0535/8049/7061/files/bg_slide1.jpg?v=1612856055"
+      alt="homepage-img"
+      className="home-img"
+    />
+    <div>
+      <img src="https://cdn.shopify.com/s/files/1/0535/8049/7061/files/slide1.1.jpg?v=1612855741" alt="slide-img" className="slide-img" />
+    </div>
+  </section>
+}
